Added LP Code
I added in Weston's LP programming and redefined a few things to match the required data types. It does take a hot second to run though because our matrices are so large. (Also it says there are redundant constraints, so maybe we can find a way to get rid of them?)